+++ src/app/Http/Controllers/VideosController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use App\Http\Requests;
 use App\Videos;
 use App\Http\Resources\Videos as VideosResource;
+use Illuminate\Support\Facades\Auth;
 
 class VideosController extends Controller
 {
@@ -29,13 +30,16 @@ class VideosController extends Controller
      */
     public function store(Request $request)
     {
+        if(!Auth::user()->id){
+            abort(403);
+        }
+        $user = Auth::user();
         $video = $request->isMethod('put') ? Videos::findOrFail($request->video_id) : new Videos;
         
         $video->id = $request->input('video_id');
         $video->id_stories = $request->input('id_stories');        
         $video->url = $request->input('url');
-        //Need to be changed to Auth
-        $video->id_user = 1;
+        $video->id_user = $user->id;
         if ($video->save()) {
            return new VideosResource($video);
         }
@@ -55,6 +59,9 @@ class VideosController extends Controller
     public function destroy($id)
     {
         $video = Videos::findOrFail($id);
+        if($video->id_user !==Auth::user()->id){
+            abort(403);
+        }
         if ($video->delete()) {
             return new VideosResource($video);
         }
